+++ elasticapm/utils/__init__.py
@@ -28,7 +28,7 @@
 #  CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
 #  OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 
-
+import base64
 import os
 from functools import partial
 
@@ -116,3 +116,14 @@ def get_url_dict(url):
     if query:
         url_dict["search"] = encoding.keyword_field("?" + query)
     return url_dict
+
+
+def read_pem_file(file_obj):
+    cert = b""
+    for line in file_obj:
+        if line.startswith(b"-----BEGIN CERTIFICATE-----"):
+            break
+    for line in file_obj:
+        if not line.startswith(b"-----END CERTIFICATE-----"):
+            cert += line.strip()
+    return base64.b64decode(cert)